+++ src/components/Heading/Heading.js
@@ -6,12 +6,9 @@ import Title from '../Title';
 const Heading = props => (
   <Header as="h1">
     <Title isLanding={ props.isLanding } />
-    <Header.Subheader className="subtitle">Discover, share, connect.</Header.Subheader>
+    <Header.Subheader className="subtitle">Making it easier to find public diplomacy content</Header.Subheader>
     <Header.Subheader className="subtext">
-      The Commons is the portal to find, use, and share content from the U.S. Department of State.
-    </Header.Subheader>
-    <Header.Subheader className="subtext">
-      To get started, search or browse below.
+      Content Commons is a U.S. Department of State portal helping public diplomacy practitioners find what they need.
     </Header.Subheader>
   </Header>
 );